changed

+++ src/hfile/fn_control.cpp
@@ -103,13 +103,13 @@ namespace rlf_hfile {
       }
 
 
-      string int_to_string( int val, size_t width ) {
-         size_t prec = 2;
+      string int_to_string( int val, int width ) {
+         int prec = 2;
          string s = rlf_hstring::toString( val, width, prec );
          return s;
       }
-      string uint_to_string( uint32_t val, size_t  width, char fill ) {
-         size_t prec = 2;
+      string uint_to_string( uint32_t val, int  width, char fill ) {
+         int prec = 2;
          string s = rlf_hstring::toString( val, width, prec, fill );
          return s;
       }
